refactor: move LearningCard to root components directory
