+++ frontend/src/components/LearningCard/ConjugationSection.module.scss
