+++ frontend/src/components/LearningCard/ConjugationSection.tsx
