+++ frontend/src/components/LearningCard/ExampleSection.module.scss
