+++ frontend/src/components/LearningCard/ExampleSection.tsx
