+++ frontend/src/components/LearningCard/LearningCard.module.scss
