+++ frontend/src/components/LearningCard/RatingButton.module.scss
