+++ frontend/src/components/LearningCard/WordSection.module.scss
