+++ frontend/src/components/LearningCard/WordSection.tsx
